fix(ordering): store money as integer instead of numeric
DECIMAL/NUMERIC columns are returned as strings by the pg driver to
preserve precision. Money.create rejected those strings because
Number.isInteger("5000") is false. Bug was dormant until GET list
endpoint triggered the first DB -> domain round-trip.

Changed unit_price and total from @Property({ type: 'decimal' }) to
@Property({ type: 'integer' }). Existing rows cast cleanly since all
values are integer-valued cents. Column type now matches the domain
invariant: Money is always integer cents, never fractional.

+++ event-ticketing/src/ordering/infrastructure/OrderEntity.ts
@@ -14,7 +14,7 @@ export class OrderEntity {
   @PrimaryKey({ type: 'string' })
   id!: string
 
-  @Property({ type: 'decimal' })
+  @Property({ type: 'integer' })
   total!: number
 
   @Property({ type: 'string' })
